+++ src/main/scala/au/com/agl/arc/metadata/MetadataSchema.scala
@@ -77,6 +77,26 @@ object MetadataSchema {
 
           t match {
 
+            case "binary" => {
+              // test keys
+              val expectedKeys = "encoding":: baseKeys
+              val invalidKeys = checkValidKeys(c)(expectedKeys)      
+
+              val encoding = getValue[String]("encoding", validValues = "base64" :: "hexadecimal" :: Nil) |> parseEncoding("encoding") _
+
+              (id, name, description, _type, nullable, nullReplacementValue, trim, nullableValues, metadata, encoding) match {
+                case (Right(id), Right(name), Right(description), Right(_type), Right(nullable), Right(nullReplacementValue), Right(trim), Right(nullableValues), Right(metadata), Right(encoding)) => {
+                  Right(BinaryColumn(id, name, description, nullable, nullReplacementValue, trim, nullableValues, encoding, metadata))
+                }
+                case _ => {
+                  val allErrors: Errors = List(id, name, description, _type, nullable, nullReplacementValue, trim, nullableValues, metadata, encoding, invalidKeys).collect{ case Left(errs) => errs }.flatten
+                  val metaName = stringOrDefault(name, "unnamed meta")
+                  val err = StageError(idx, metaName, c.origin.lineNumber, allErrors)
+                  Left(err :: Nil)
+                }
+              }  
+            }              
+
             case "boolean" => {
               // test keys
               val expectedKeys = "trueValues" :: "falseValues" :: baseKeys